requirements.txt
fastmcp>=2.0.0
discord.py>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
